Fix importmap usage in tests
