+++ package.json
@@ -13,7 +13,7 @@
     "url": "https://github.com/jsenv/core"
   },
   "engines": {
-    "node": ">=18.5.0"
+    "node": ">=20.8.0"
   },
   "publishConfig": {
     "access": "public"
