+++ packages/related/test/src/runtime_node/node_worker_thread.js
@@ -65,9 +65,7 @@ export const nodeWorkerThread = ({
       if (importMap) {
         env.IMPORT_MAP = JSON.stringify(importMap);
         env.IMPORT_MAP_BASE_URL = rootDirectoryUrl;
-        commandLineOptions.push(
-          `--experimental-loader=${IMPORTMAP_NODE_LOADER_FILE_URL}`,
-        );
+        commandLineOptions.push(`--import=${IMPORTMAP_NODE_LOADER_FILE_URL}`);
         commandLineOptions.push(
           `--require=${fileURLToPath(NO_EXPERIMENTAL_WARNING_FILE_URL)}`,
         );
